backend/order/pyproject.toml
[project]
name = "order"
version = "0.1.0"
description = ""
authors = [
    {name = "vansh",email = "vanshpatel1299@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"  # Adjusted Python version requirement
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "setuptools (>=75.8.0,<76.0.0)",
    "redis (>=4.0,<5.0)",
    "aio-pika (>=9.5.4,<10.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
